requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pandas>=2.3.0
openpyxl>=3.1.2
python-multipart>=0.0.6
pydantic>=2.5.0
python-dateutil>=2.9.0
pytz>=2025.1
tzdata>=2025.1

